// lib/types.ts
export interface components {
    schemas: {
// ...
        MusicResponse: {
            /** Id */
            id: string;
            /** Project Id */
            project_id: string;
            /** User Id */
            user_id: string;
            /** User Name */
            user_name: string;
            /** User Email */
            user_email: string;
            /** Type */
            type: string;
            /** Task Id */
            task_id: string;
            /** Conversion Id */
            conversion_id: string;
            /** Status */
            status: string;
            /** Prompt */
            prompt?: string | null;
            /** Music Style */
            music_style?: string | null;
            /** Title */
            title?: string | null;
            /** Duration */
            duration?: number | null;
            /** Audio Url */
            audio_url?: string | null;
            /** Album Cover Path */
            album_cover_path?: string | null;
            /** Generated Lyrics */
            generated_lyrics?: string | null;
            /** Is Cloned */
            is_cloned?: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Updated At */
            updated_at?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}